re-did mobile build from ground up
