+++ Selu383.SP24.Mobile/.env
@@ -1,1 +1,1 @@
-EXPO_PACKAGER_PROXY_URL= "http://2c69-147-174-75-100.ngrok-free.app"
+EXPO_PACKAGER_PROXY_URL=http://1eb5-2600-6c80-300-2f-484c-933a-f66c-7f52.ngrok-free.app
