+++ Selu383.SP24.Mobile/App.js
@@ -1,54 +1,20 @@
-// App.js
-import React from 'react';
-import { NavigationContainer } from '@react-navigation/native';
-import { createDrawerNavigator } from '@react-navigation/drawer';
-import LoginScreen from './src/screens/LoginScreen';
-import HomeScreen from "./src/screens/HomeScreen";
+import { StatusBar } from 'expo-status-bar';
+import { StyleSheet, Text, View } from 'react-native';
 
-const Drawer = createDrawerNavigator();
-
-export default function App(){
-    return (
-    
-    <NavigationContainer>
-            <Drawer.Navigator initialRouteName="Login" screenOptions={{drawerActiveTintColor: "#1261E6"}}>
-                <Drawer.Screen
-                    name="Home"
-                    component={HomeScreen}
-                />
-                <Drawer.Screen
-                    name="Book A Room"
-                    component={HomeScreen}
-                />
-                <Drawer.Screen
-                    name="Reservations"
-                    component={HomeScreen}
-                />
-                <Drawer.Screen
-                    name="Locations"
-                    component={HomeScreen}
-                />
-                <Drawer.Screen
-                    name="Services"
-                    component={HomeScreen}
-                />
-                <Drawer.Screen
-                    name="Settings"
-                    component={HomeScreen}
-                />
-                <Drawer.Screen
-                    name="About"
-                    component={HomeScreen}
-                />
-                <Drawer.Screen
-                    name="Contact Us"
-                    component={HomeScreen}
-                />
-                <Drawer.Screen
-                    name="Login"
-                    component={LoginScreen}
-                />
-            </Drawer.Navigator>
-    </NavigationContainer>
-    );
+export default function App() {
+  return (
+    <View style={styles.container}>
+      <Text>Open up App.js to start working on your app!</Text>
+      <StatusBar style="auto" />
+    </View>
+  );
 }
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#fff',
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+});
