+++ Selu383.SP24.Mobile/app.json
@@ -29,7 +29,7 @@
     },
     "extra": {
       "eas": {
-        "projectId": "f5af501c-6d12-404d-9ebe-10020a5c80f1"
+        "projectId": "c4ce24ea-965b-4aa8-b7a2-acd89dc7eb68"
       }
     },
     "owner": "leblanc.gavin"
