--- Selu383.SP24.Mobile/src/screens/HomeScreen.js
@@ -1,14 +1,0 @@
-import React from 'react';
-import { View, StyleSheet } from 'react-native';
-
-const HomeScreen = () => {
-  
-};
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-  },
-});
-
-export default HomeScreen;
